[KBM] Move resources from the Common project to the Editor. (#10844)
