+++ src/modules/keyboardmanager/KeyboardManagerEditor/EditKeyboardWindow.cpp
@@ -11,9 +11,8 @@
 #include <KeyboardManagerConstants.h>
 #include <KeyboardManagerState.h>
 
-#include <dll/Generated Files/resource.h>
-
 #include "EditKeyboardWindow.h"
+#include "ErrorTypes.h"
 #include "SingleKeyRemapControl.h"
 #include "KeyDropDownControl.h"
 #include "XamlBridge.h"
