+++ src/modules/keyboardmanager/KeyboardManagerEditor/KeyboardManagerEditor.vcxproj.filters
@@ -63,7 +63,7 @@
     <ClInclude Include="resource.base.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Constants.h">
+    <ClInclude Include="KeyboardManagerEditorStrings.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="Generated Files\resource.h">
@@ -108,6 +108,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="XamlBridge.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="KeyboardManagerEditorStrings.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
