+++ src/modules/keyboardmanager/KeyboardManagerEditor/KeyboardManagerEditorStrings.cpp
@@ -1,0 +1,50 @@
+#include "pch.h"
+#include "KeyboardManagerEditorStrings.h"
+
+// Function to return the error message
+winrt::hstring KeyboardManagerEditorStrings::GetErrorMessage(KeyboardManagerHelper::ErrorType errorType)
+{
+    using namespace KeyboardManagerHelper;
+
+    switch (errorType)
+    {
+    case ErrorType::NoError:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_REMAPSUCCESSFUL).c_str();
+    case ErrorType::SameKeyPreviouslyMapped:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_SAMEKEYPREVIOUSLYMAPPED).c_str();
+    case ErrorType::MapToSameKey:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_MAPPEDTOSAMEKEY).c_str();
+    case ErrorType::ConflictingModifierKey:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_CONFLICTINGMODIFIERKEY).c_str();
+    case ErrorType::SameShortcutPreviouslyMapped:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_SAMESHORTCUTPREVIOUSLYMAPPED).c_str();
+    case ErrorType::MapToSameShortcut:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_MAPTOSAMESHORTCUT).c_str();
+    case ErrorType::ConflictingModifierShortcut:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_CONFLICTINGMODIFIERSHORTCUT).c_str();
+    case ErrorType::WinL:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_WINL).c_str();
+    case ErrorType::CtrlAltDel:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_CTRLALTDEL).c_str();
+    case ErrorType::RemapUnsuccessful:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_REMAPUNSUCCESSFUL).c_str();
+    case ErrorType::SaveFailed:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_SAVEFAILED).c_str();
+    case ErrorType::ShortcutStartWithModifier:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_SHORTCUTSTARTWITHMODIFIER).c_str();
+    case ErrorType::ShortcutCannotHaveRepeatedModifier:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_SHORTCUTNOREPEATEDMODIFIER).c_str();
+    case ErrorType::ShortcutAtleast2Keys:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_SHORTCUTATLEAST2KEYS).c_str();
+    case ErrorType::ShortcutOneActionKey:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_SHORTCUTONEACTIONKEY).c_str();
+    case ErrorType::ShortcutNotMoreThanOneActionKey:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_SHORTCUTMAXONEACTIONKEY).c_str();
+    case ErrorType::ShortcutMaxShortcutSizeOneActionKey:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_MAXSHORTCUTSIZE).c_str();
+    case ErrorType::ShortcutDisableAsActionKey:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_DISABLEASACTIONKEY).c_str();
+    default:
+        return GET_RESOURCE_STRING(IDS_ERRORMESSAGE_DEFAULT).c_str();
+    }
+}
