+++ src/modules/keyboardmanager/KeyboardManagerEditor/KeyboardManagerEditorStrings.h
@@ -2,8 +2,13 @@
 
 #include "pch.h"
 
-namespace KeyboardManagerConstants
+#include <ErrorTypes.h>
+
+namespace KeyboardManagerEditorStrings
 {
     // String constant for the default app name in Remap shortcuts
     inline const std::wstring DefaultAppName = GET_RESOURCE_STRING(IDS_EDITSHORTCUTS_ALLAPPS);
+        
+    // Function to return the error message
+    winrt::hstring GetErrorMessage(KeyboardManagerHelper::ErrorType errorType);
 }
