+++ src/modules/keyboardmanager/KeyboardManagerEditor/Resources.resx
@@ -188,6 +188,69 @@
   <data name="EditShortcuts_InfoExample" xml:space="preserve">
     <value>For example, if you want to press "Ctrl+C" and get "Alt" only on Microsoft Edge, "Ctrl+C" would be your "Shortcut" column, the key "Alt" would be your "Mapped To" column, and "MSEdge" would be your "Target App" column. If no target app is entered, it will apply globally. The name must be the process name and not the app name.</value>
   </data>
+  <data name="ErrorMessage_RemapSuccessful" xml:space="preserve">
+    <value>Remapping successful</value>
+  </data>
+  <data name="ErrorMessage_RemapUnsuccessful" xml:space="preserve">
+    <value>Some remappings were not applied</value>
+  </data>
+  <data name="ErrorMessage_SameKeyPreviouslyMapped" xml:space="preserve">
+    <value>Cannot remap a key more than once</value>
+    <comment>Key on a keyboard</comment>
+  </data>
+  <data name="ErrorMessage_MappedToSameKey" xml:space="preserve">
+    <value>Cannot remap a key to itself</value>
+    <comment>Key on a keyboard</comment>
+  </data>
+  <data name="ErrorMessage_ConflictingModifierKey" xml:space="preserve">
+    <value>Cannot remap this key as it conflicts with another remapped key</value>
+    <comment>Key on a keyboard</comment>
+  </data>
+  <data name="ErrorMessage_SameShortcutPreviouslyMapped" xml:space="preserve">
+    <value>Cannot remap a shortcut more than once for the same target app</value>
+  </data>
+  <data name="ErrorMessage_MapToSameShortcut" xml:space="preserve">
+    <value>Cannot remap a shortcut to itself</value>
+  </data>
+  <data name="ErrorMessage_ConflictingModifierShortcut" xml:space="preserve">
+    <value>Cannot remap this shortcut as it conflicts with another remapped shortcut</value>
+  </data>
+  <data name="ErrorMessage_WinL" xml:space="preserve">
+    <value>Cannot remap from/to Win L</value>
+    <comment>Win refers to Windows as in the OS</comment>
+  </data>
+  <data name="ErrorMessage_CtrlAltDel" xml:space="preserve">
+    <value>Cannot remap from/to Ctrl Alt Del</value>
+  </data>
+  <data name="ErrorMessage_SaveFailed" xml:space="preserve">
+    <value>Failed to save the remappings</value>
+  </data>
+  <data name="ErrorMessage_ShortcutStartWithModifier" xml:space="preserve">
+    <value>Shortcut must start with a modifier key</value>
+    <comment>Key on a keyboard</comment>
+  </data>
+  <data name="ErrorMessage_ShortcutNoRepeatedModifier" xml:space="preserve">
+    <value>Shortcut cannot contain a repeated modifier</value>
+  </data>
+  <data name="ErrorMessage_ShortcutAtleast2Keys" xml:space="preserve">
+    <value>Shortcut must have atleast 2 keys</value>
+    <comment>Key on a keyboard</comment>
+  </data>
+  <data name="ErrorMessage_ShortcutOneActionKey" xml:space="preserve">
+    <value>Shortcut must contain an action key</value>
+    <comment>Key on a keyboard</comment>
+  </data>
+  <data name="ErrorMessage_ShortcutMaxOneActionKey" xml:space="preserve">
+    <value>Shortcut cannot have more than one action key</value>
+    <comment>Key on a keyboard</comment>
+  </data>
+  <data name="ErrorMessage_MaxShortcutSize" xml:space="preserve">
+    <value>Shortcuts can only have up to 2 modifier keys</value>
+    <comment>Key on a keyboard</comment>
+  </data>
+  <data name="ErrorMessage_Default" xml:space="preserve">
+    <value>Unexpected error</value>
+  </data>
   <data name="Type_Button" xml:space="preserve">
     <value>Type</value>
     <comment>As in type a key</comment>
