+++ src/modules/keyboardmanager/common/Helpers.h
@@ -29,29 +29,6 @@ namespace KeyboardManagerHelper
         Action
     };
 
-    // Type to store codes for different errors
-    enum class ErrorType
-    {
-        NoError,
-        SameKeyPreviouslyMapped,
-        MapToSameKey,
-        ConflictingModifierKey,
-        SameShortcutPreviouslyMapped,
-        MapToSameShortcut,
-        ConflictingModifierShortcut,
-        WinL,
-        CtrlAltDel,
-        RemapUnsuccessful,
-        SaveFailed,
-        ShortcutStartWithModifier,
-        ShortcutCannotHaveRepeatedModifier,
-        ShortcutAtleast2Keys,
-        ShortcutOneActionKey,
-        ShortcutNotMoreThanOneActionKey,
-        ShortcutMaxShortcutSizeOneActionKey,
-        ShortcutDisableAsActionKey
-    };
-
     // Enum type to store possible decision for input in the low level hook
     enum class KeyboardHookDecision
     {
@@ -77,9 +54,6 @@ namespace KeyboardManagerHelper
 
     // Function to check if two keys are equal or cover the same set of keys. Return value depends on type of overlap
     ErrorType DoKeysOverlap(DWORD first, DWORD second);
-
-    // Function to return the error message
-    winrt::hstring GetErrorMessage(ErrorType errorType);
 
     // Function to return the list of key name in the order for the drop down based on the key codes
     winrt::Windows::Foundation::Collections::IVector<winrt::Windows::Foundation::IInspectable> ToBoxValue(const std::vector<std::pair<DWORD,std::wstring>>& list);
