+++ src/modules/keyboardmanager/common/Shortcut.cpp
@@ -2,6 +2,7 @@
 #include "Shortcut.h"
 #include <common/interop/keyboard_layout.h>
 #include <common/interop/shared_constants.h>
+#include "ErrorTypes.h"
 #include "Helpers.h"
 #include "InputInterface.h"
